+++ sdk/nodejs/index.ts
@@ -5,6 +5,7 @@
 export * from "./dashboard";
 export * from "./dashboardList";
 export * from "./downtime";
+export * from "./getDashboard";
 export * from "./getDashboardList";
 export * from "./getIpRanges";
 export * from "./getMonitor";